remove Edge, old includes, EdgeListReaders, GPUCSR

+++ include/pangolin/algorithm/tc_vertex_warp_bitvector.cuh
@@ -131,7 +131,7 @@ __global__ void __launch_bounds__(BLOCK_DIM_X) tc_vertex_warp_bitvector_kernel(
 
     } else { // use another approach
       if (0 == threadIdx.x) {
-        printf("AHH Block %d row %d wont fit, needs %lu\n", blockIdx.x, warpRow, rowSz);
+        printf("AHH Block %d row %lu wont fit, needs %lu\n", blockIdx.x, uint64_t(warpRow), uint64_t(rowSz));
       }
     }
   }
